+++ NEWS.md
@@ -7,6 +7,9 @@
 * Add support for labeling tiles
 * Added new tests for the `ggRelatednessMatrix` function.
 * Added new phenotype plotting function `ggPhenotypebyDegree` to visualize phenotypes by degree of relatedness.
+* Added new vignette to demonstrate `ggPhenotypebyDegree`.
+* Add default config function to ensure consistent configuration across functions.
+
 
 # ggpedigree 0.4.1
 ## Status: Submitted to CRAN